pyproject.toml
[tool.poetry]
name = "fastapi-blog-api"
version = "0.1.0"
description = ""
authors = ["penkhasoveg <pen.egor2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.95.1" }
uvicorn = "^0.22.0"
typer = {extras = ["all"], version = "^0.9.0"}
sqlalchemy = "^2.0.12"
alembic = "^1.10.4"
asyncpg = "^0.27.0"
pyhumps = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
